third_party/blink/web_tests/external/wpt/dom/nodes/NodeList-static-length-getter-tampered-3-expected.txt

This is a testharness.js-based test.
FAIL NodeList (static collection) "length" getter tampered assert_equals: expected -1 but got 50
Harness: the test ran to completion.
